Move checks into separate packages.
This is a rather beefy changeset. At the heart of the effort is
a desire to make each check live in its own package to prevent
polluting package level types into the other check packages.

Because there is some level of code that is shared amongst most
of the packages, a new package was created for this code:

github.com/dcos/dcos-checks/common

Since the `init` functions are not automatically applied for each
of the checks, the subcommands.go file now loads each check
manually using a function exported from each package:

func Add(root *cobra.Command)

The package implementation of this function adds the subcommand
to the root.

Other changes that came along for the ride include:

- move types into their own files where appropriate
- organized files by const -> type -> var
- moved fixture folder to be a child of the `ip` package
- creates `constants` package for shared constants
- unexported types in all check packages
- only Add() method exported in each package

I decided that due to the level of refactoring in this changeset
that the "Reviewer First" method realistically only applies to
this commit message.